backend/requirements.txt
click==7.1.2
Flask==1.1.2
Flask-Cors==3.0.9
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
pkg-resources==0.0.0
psycopg2==2.8.6
python-dotenv==0.15.0
six==1.15.0
SQLAlchemy==1.3.20
SQLAlchemy-Utils==0.36.8
Werkzeug==1.0.1
